refactor: move DeferredResultRenderer methods to ExtensionSocketServer
