+++ tests/modes/apps/extensions/test_ExtensionSocketController.py
@@ -2,7 +2,6 @@ from unittest import mock
 
 import pytest
 
-from ulauncher.api.shared.query import Query
 from ulauncher.modes.extensions.ExtensionSocketController import ExtensionSocketController
 
 TEST_EXT_ID = "com.example.test-ext-id"
@@ -12,10 +11,6 @@ class TestExtensionSocketController:
     @pytest.fixture
     def controllers(self):
         return {}
-
-    @pytest.fixture(autouse=True)
-    def result_renderer(self, mocker):
-        return mocker.patch("ulauncher.modes.extensions.ExtensionSocketController.DeferredResultRenderer").return_value
 
     @pytest.fixture(autouse=True)
     def ec_extension_finder(self, mocker):
@@ -46,22 +41,7 @@ class TestExtensionSocketController:
         controller.trigger_event(event)
         controller.framer.send.assert_called_with(event)
 
-    def test_handle_query__KeywordQueryEvent__is_sent_with_query(self, controller, result_renderer):
-        query = Query("def ulauncher")
-        assert controller.handle_query(query) == result_renderer.handle_event.return_value
-        keywordEvent = controller.framer.send.call_args_list[1][0][0]
-        assert keywordEvent.get("type") == "event:input_trigger"
-        assert keywordEvent.get("args", [])[0] == "ulauncher"
-        result_renderer.handle_event.assert_called_with(keywordEvent, controller)
-
     def test_handle_response__unsupported_data_type__exception_raised(self, controller):
         controller.data = {}
         with pytest.raises(TypeError):
             controller.handle_response(controller.framer, object())
-
-    def test_handle_response__is_called(self, controller, result_renderer):
-        response = {"event": mock.Mock(), "action": {}}
-
-        controller.handle_response(controller.framer, response)
-
-        result_renderer.handle_response.assert_called_with(response, controller)
